Fix 配色系をConfigファイルに切り分け
