+++ src/features/pictures/components/canvas/ColorPicker.tsx
@@ -1,54 +1,7 @@
 import { EraserIcon } from '@radix-ui/react-icons'
 import React, { memo } from 'react'
 
-type ColorType = {
-  name: string
-  class: string
-  color: string
-}
-
-const Colors: ColorType[] = [
-  {
-    name: '黒',
-    class: 'bg-black',
-    color: '#000000',
-  },
-  {
-    name: '青',
-    class: 'bg-blue-500',
-    color: '#3B82F6',
-  },
-  {
-    name: '赤',
-    class: 'bg-red-500',
-    color: '#EF4444',
-  },
-  {
-    name: '黄',
-    class: 'bg-yellow-500',
-    color: '#EAB308',
-  },
-  {
-    name: '緑',
-    class: 'bg-green-500',
-    color: '#22C55E',
-  },
-  {
-    name: '紫',
-    class: 'bg-purple-500',
-    color: '#A855F7',
-  },
-  {
-    name: '橙',
-    class: 'bg-orange-500',
-    color: '#F97316',
-  },
-  {
-    name: '消しゴム',
-    class: '',
-    color: '#FFFFFF',
-  },
-]
+import { Colors } from '@/utils/config/colors'
 
 type ColorPickerProps = {
   setLineWidth: React.Dispatch<React.SetStateAction<number>>
